Add Virtual host details Add xdebug ini settings for remote Add swap directory details Bug fix to shell installing

+++ provision/ssh-keygen.sh
@@ -13,17 +13,11 @@ chmod 644 '/home/vagrant/.ssh/id_rsa.pub'
 chmod 600 '/home/vagrant/.ssh/id_rsa'
 chmod 600 '/home/vagrant/.ssh/authorized_keys'
 
-cat ~/.ssh/id_rsa.pub | tee ~/.ssh/authorized_keys
+if [ -f ~/.ssh/id_rsa.pub ]; then
+  cat ~/.ssh/id_rsa.pub | tee ~/.ssh/authorized_keys
+fi
 
 # Know our hosts
 ssh-keyscan -H 104.236.178.120 >> ~/.ssh/known_hosts
 ssh-keyscan -H github.com >> ~/.ssh/known_hosts
 ssh-keyscan -H bitbucket.org >> ~/.ssh/known_hosts
-
-cp "${VAGRANT_CORE_FOLDER}/.bash_aliases" '/home/vagrant/.bash_aliases'
-cp "${VAGRANT_CORE_FOLDER}/.bash_git" '/home/vagrant/.bash_git'
-
-chown vagrant:vagrant /home/vagrant/.bash_aliases
-chown vagrant:vagrant /home/vagrant/.bash_git
-
-source /home/vagrant/.bashrc
